add: initial version of keyboard layout
In this commit, the keyboard layout finally gets displayed correctly.
Every button normally found on a keyboard is placed in a grid and then
shown in the window.
We probably want to change this to not use a grid in the future, because
keyboards normally have their rows slightly offset from each other